+++ src/components/pages/dpm/DpmCalcPage.tsx
@@ -1,14 +1,8 @@
-import { useState, useCallback, useMemo } from 'react';
+import { useState, useCallback } from 'react';
 import Box from '@mui/material/Box';
 import Stack from '@mui/material/Stack';
 import Paper from '@mui/material/Paper';
 import Typography from '@mui/material/Typography';
-import Table from '@mui/material/Table';
-import TableBody from '@mui/material/TableBody';
-import TableCell from '@mui/material/TableCell';
-import TableContainer from '@mui/material/TableContainer';
-import TableHead from '@mui/material/TableHead';
-import TableRow from '@mui/material/TableRow';
 import { NavigationBar } from '../../navigation/NavigationBar';
 import { PageContent } from '../../pageStructure/PageContent';
 import { createWeaponBaseProperties, createWeaponEnhancementProperties, createTargetProperties } from './utils/dpmCalcInputUtils';
@@ -17,9 +11,7 @@ import { WeaponBasePropertyForm } from './WeaponBasePropertyForm';
 import { PageFooter } from '../../pageStructure/PageFooter';
 import { usePropertyTabs } from './hooks/usePropertyTabs';
 import { PropertyTabs } from './PropertyTabs';
-import { INumericOutputProperty, IOutputProperties, OutputPropertyId } from './types/IOutputProperty';
-import { createOutputProperties, createOutputPropertiesForTabs, dependsOn } from './utils/dpmCalcOutputUtils';
-import { LabeledList } from '../../list/LabeledList';
+import { ComputedProperties } from './ComputedProperties';
 
 export const DpmCalcPage = () => {
     const [weaponBaseProperties, setWeaponBaseProperties] = useState<IWeaponBaseProperties>(() => createWeaponBaseProperties());
@@ -60,15 +52,6 @@ export const DpmCalcPage = () => {
             [newInputProperty.id]: newInputProperty,
         }));
     }, []);
-
-    const baseOutputProperties = useMemo<Readonly<IOutputProperties>>(() => createOutputProperties(), []);
-
-    const computedOutputProperties = useMemo<Record<string, Record<string, IOutputProperties>>>(() => createOutputPropertiesForTabs({
-        weaponBaseProperties,
-        enhancementTabs,
-        attackTargetTabs,
-        baseOutputProperties,
-    }), [weaponBaseProperties, enhancementTabs, attackTargetTabs, baseOutputProperties]);
 
     return (
         <>
@@ -137,146 +120,11 @@ export const DpmCalcPage = () => {
                             </Typography>
                             <Paper>
                                 <Box p={1}>
-                                    {enhancementTabs.length === 1 && attackTargetTabs.length === 1 && (
-                                        <LabeledList
-                                            rows={
-                                                Object.values(baseOutputProperties)
-                                                    .map(baseOutputProperty => ({
-                                                        key: baseOutputProperty.id,
-                                                        label: baseOutputProperty.label,
-                                                        value: (() => {
-                                                            const computedProperty = computedOutputProperties[enhancementTabs[0].id]?.[attackTargetTabs[0].id]?.[baseOutputProperty.id as OutputPropertyId] ?? null; 
-                                                            if (!computedProperty) {
-                                                                return '-';
-                                                            }
-                                                            if (computedProperty.type === 'numeric') {
-                                                                return (computedProperty as INumericOutputProperty).value ?? '-';
-                                                            }
-                                                            return '?';
-                                                        })(),
-                                                    }))
-                                            }
-                                        />
-                                    )}
-                                    {(enhancementTabs.length > 1 || attackTargetTabs.length > 1) && Object.values(baseOutputProperties).map(baseOutputProperty => {
-                                        const dependsOnWeaponEnhancement = dependsOn(baseOutputProperty, 'weaponEnhancementProperties', baseOutputProperties);
-                                        const dependsOnTarget = dependsOn(baseOutputProperty, 'targetProperties', baseOutputProperties);
-                                        const showColumnsForEnhancement = dependsOnWeaponEnhancement && enhancementTabs.length > 1;
-                                        const showRowsForTargets = showColumnsForEnhancement && dependsOnTarget && attackTargetTabs.length > 1;
-                                        const showColumnsForTargets = !showColumnsForEnhancement && dependsOnTarget && attackTargetTabs.length > 1;
-
-                                        return (
-                                            <Box key={baseOutputProperty.id} pb={6}>
-                                                <Typography variant="body1" gutterBottom={true}>{baseOutputProperty.label}</Typography>
-                                                {showColumnsForEnhancement && (
-                                                    <Box pl={2}>
-                                                        <TableContainer component={Paper} sx={{ width: 'fit-content' }}>
-                                                            <Table size="small" sx={{ width: 'fit-content' }}>
-                                                                <TableHead>
-                                                                    <TableRow>
-                                                                        {showRowsForTargets && (
-                                                                            <TableCell>{''}</TableCell>
-                                                                        )}
-                                                                        {enhancementTabs.map((enhancementTab, index) => (
-                                                                            <TableCell key={enhancementTab.id}>
-                                                                                <Typography variant="body2" color="text.secondary">
-                                                                                    {enhancementTab.name || enhancementTab.defaultName(index)}
-                                                                                </Typography>
-                                                                            </TableCell>
-                                                                        ))}
-                                                                    </TableRow>
-                                                                </TableHead>
-                                                                <TableBody>
-                                                                    {(showRowsForTargets ? attackTargetTabs : [attackTargetTabs[0]]).map((attackTargetTab, index) => (
-                                                                        <TableRow key={attackTargetTab.id}>
-                                                                            {showRowsForTargets && (
-                                                                                <TableCell>
-                                                                                    <Typography variant="body2" color="text.secondary">
-                                                                                        {attackTargetTab.name || attackTargetTab.defaultName(index)}
-                                                                                    </Typography>
-                                                                                </TableCell>
-                                                                            )}
-                                                                            {enhancementTabs.map(enhancementTab => (
-                                                                                <TableCell key={enhancementTab.id}>
-                                                                                    <Typography variant="body2" color="text.secondary">
-                                                                                        {(() => {
-                                                                                            const computedProperty = computedOutputProperties[enhancementTab.id]?.[attackTargetTab.id]?.[baseOutputProperty.id as OutputPropertyId] ?? null; 
-                                                                                            if (!computedProperty) {
-                                                                                                return '-';
-                                                                                            }
-                                                                                            if (computedProperty.type === 'numeric') {
-                                                                                                return (computedProperty as INumericOutputProperty).value ?? '-';
-                                                                                            }
-                                                                                            return '?';
-                                                                                        })()}
-                                                                                    </Typography>
-                                                                                </TableCell>
-                                                                            ))}
-                                                                        </TableRow>
-                                                                    ))}
-                                                                </TableBody>
-                                                            </Table>
-                                                        </TableContainer>
-                                                    </Box>
-                                                )}
-                                                {!showColumnsForEnhancement && showColumnsForTargets && (
-                                                    <Box pl={2}>
-                                                        <TableContainer component={Paper} sx={{ width: 'fit-content' }}>
-                                                            <Table size="small" sx={{ width: 'fit-content' }}>
-                                                                <TableHead>
-                                                                    <TableRow>
-                                                                        {attackTargetTabs.map((attackTargetTab, index) => (
-                                                                            <TableCell key={attackTargetTab.id}>
-                                                                                <Typography variant="body2" color="text.secondary">
-                                                                                    {attackTargetTab.name || attackTargetTab.defaultName(index)}
-                                                                                </Typography>
-                                                                            </TableCell>
-                                                                        ))}
-                                                                    </TableRow>
-                                                                </TableHead>
-                                                                <TableBody>
-                                                                    <TableRow>
-                                                                        {attackTargetTabs.map((attackTargetTab, index) => (
-                                                                            <TableCell key={attackTargetTab.id}>
-                                                                                <Typography variant="body2" color="text.secondary">
-                                                                                    {(() => {
-                                                                                        const computedProperty = computedOutputProperties[enhancementTabs[0].id]?.[attackTargetTab.id]?.[baseOutputProperty.id as OutputPropertyId] ?? null; 
-                                                                                        if (!computedProperty) {
-                                                                                            return '-';
-                                                                                        }
-                                                                                        if (computedProperty.type === 'numeric') {
-                                                                                            return (computedProperty as INumericOutputProperty).value ?? '-';
-                                                                                        }
-                                                                                        return '?';
-                                                                                    })()}
-                                                                                </Typography>
-                                                                            </TableCell>
-                                                                        ))}
-                                                                    </TableRow>
-                                                                </TableBody>
-                                                            </Table>
-                                                        </TableContainer>
-                                                    </Box>
-                                                )}
-                                                {!showColumnsForEnhancement && !showColumnsForTargets && (
-                                                    <Box pl={2}>
-                                                        <Typography variant="body2" color="text.secondary">
-                                                            {(() => {
-                                                                const computedProperty = computedOutputProperties[enhancementTabs[0].id]?.[attackTargetTabs[0].id]?.[baseOutputProperty.id as OutputPropertyId] ?? null; 
-                                                                if (!computedProperty) {
-                                                                    return '-';
-                                                                }
-                                                                if (computedProperty.type === 'numeric') {
-                                                                    return (computedProperty as INumericOutputProperty).value ?? '-';
-                                                                }
-                                                                return '?';
-                                                            })()}
-                                                        </Typography>
-                                                    </Box>
-                                                )}
-                                            </Box>
-                                        );
-                                    })}
+                                    <ComputedProperties
+                                        weaponBaseProperties={weaponBaseProperties}
+                                        enhancementTabs={enhancementTabs}
+                                        attackTargetTabs={attackTargetTabs}
+                                    />
                                 </Box>
                             </Paper>
                         </Stack>
