+++ src/components/pages/dpm/utils/unitAdornmentUtils.ts
@@ -1,0 +1,12 @@
+import { Unit } from '../types/Unit';
+
+export function getAdornmentForUnit(unit: Unit) {
+    switch (unit) {
+        case Unit.PERCENTAGE:
+            return '%';
+        case Unit.SECONDS:
+            return '秒';
+        case Unit.DPM:
+            return '/分';
+    }
+}
